move delta logic to own module
